Restructured AppSettings
