+++ lib/app/modules/home/views/add_task_bottom_sheet.dart
@@ -12,7 +12,7 @@ import 'package:taskwarrior/app/modules/home/controllers/widget.controller.dart'
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/taskfunctions/taskparser.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 class AddTaskBottomSheet extends StatelessWidget {
   final HomeController homeController;
