+++ lib/app/modules/home/views/home_page_nav_drawer_menu_item.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
