+++ lib/app/modules/manageTaskServer/views/manage_task_server_page_app_bar.dart
@@ -11,7 +11,7 @@ import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/constants/utilites.dart';
 
 import 'package:taskwarrior/app/utils/gen/fonts.gen.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../controllers/manage_task_server_controller.dart';
