+++ lib/app/modules/profile/views/deleteprofiledialog.dart
@@ -3,7 +3,7 @@ import 'package:get/get.dart';
 import 'package:taskwarrior/app/modules/splash/controllers/splash_controller.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
 import 'package:taskwarrior/app/utils/constants/utilites.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 class DeleteProfileDialog extends StatelessWidget {
   const DeleteProfileDialog({
