+++ lib/app/modules/profile/views/selectprofile.dart
@@ -5,7 +5,7 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:taskwarrior/app/modules/home/controllers/home_controller.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 class SelectProfile extends StatelessWidget {
   const SelectProfile(
