+++ lib/app/modules/reports/views/burn_down_monthly.dart
@@ -7,7 +7,7 @@ import 'package:taskwarrior/app/modules/reports/views/common_chart_indicator.dar
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/gen/fonts.gen.dart';
 import 'package:taskwarrior/app/utils/language/sentence_manager.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 class BurnDownMonthly extends StatelessWidget {
   final ReportsController reportsController;
