+++ lib/app/modules/reports/views/reports_view.dart
@@ -8,7 +8,7 @@ import 'package:taskwarrior/app/modules/reports/views/burn_down_weekly.dart';
 import 'package:taskwarrior/app/utils/constants/constants.dart';
 import 'package:taskwarrior/app/utils/gen/fonts.gen.dart';
 import 'package:taskwarrior/app/utils/language/sentence_manager.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 import '../controllers/reports_controller.dart';
