+++ lib/app/modules/settings/controllers/settings_controller.dart
@@ -13,7 +13,7 @@ import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/constants/utilites.dart';
 
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 import 'package:path/path.dart' as path;
 
 import 'package:taskwarrior/app/modules/splash/controllers/splash_controller.dart';
