+++ lib/app/modules/settings/views/settings_page_select_directory_list_tile.dart
@@ -9,7 +9,7 @@ import 'package:taskwarrior/app/modules/splash/controllers/splash_controller.dar
 import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/constants/utilites.dart';
 import 'package:taskwarrior/app/utils/language/sentence_manager.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 import '../controllers/settings_controller.dart';
