+++ lib/app/modules/settings/views/settings_view.dart
@@ -6,7 +6,7 @@ import 'package:get/get.dart';
 import 'package:taskwarrior/app/modules/settings/views/settings_page_app_bar.dart';
 import 'package:taskwarrior/app/modules/settings/views/settings_page_body.dart';
 import 'package:taskwarrior/app/utils/constants/taskwarrior_colors.dart';
-import 'package:taskwarrior/app/utils/app_settings.dart';
+import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
 
 import '../controllers/settings_controller.dart';
