+++ lib/app/utils/app_settings/selected_language.dart
@@ -1,0 +1,19 @@
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:taskwarrior/app/utils/language/supported_language.dart';
+
+class SelectedLanguage {
+  static SharedPreferences? _preferences;
+
+  static Future init() async {
+    _preferences = await SharedPreferences.getInstance();
+  }
+
+  static Future saveSelectedLanguage(SupportedLanguage language) async {
+    await _preferences?.setString('_selectedLanguage', language.languageCode);
+  }
+
+  static SupportedLanguage? getSelectedLanguage() {
+    String? languageCode = _preferences?.getString('_selectedLanguage');
+    return SupportedLanguageExtension.fromCode(languageCode);
+  }
+}
