+++ lib/app/utils/app_settings/selected_theme.dart
@@ -1,0 +1,17 @@
+import 'package:shared_preferences/shared_preferences.dart';
+
+class SelectedTheme {
+  static SharedPreferences? _preferences;
+
+  static Future init() async {
+    _preferences = await SharedPreferences.getInstance();
+  }
+
+  static Future saveMode(bool mode) async {
+    await _preferences?.setBool('_isDark', mode);
+  }
+
+  static bool? getMode() {
+    return _preferences?.getBool('_isDark');
+  }
+}
