+++ lib/app/utils/tour/details_page_tour.dart
